Refactor Javascript modules
