+++ flower/static/src/css/flower.scss
@@ -1,4 +1,4 @@
-@import "~bootstrap/scss/bootstrap";
+@import "../../../../node_modules/bootstrap/scss/bootstrap";
 @import "npm:datatables.net-bs5/css/dataTables.bootstrap5.css";
 
 body {
