+++ flower/static/src/js/http.js
@@ -1,0 +1,3 @@
+export const JSON_HTTP_HEADERS = {
+    "Content-Type": "application/json",
+};
